tvm-bench: update util.py and mobilnet benchmarks
Updating util.py and mobilenet benchmarks to to include def extract(path) and def load_test_image for better reusibility.

utils.py: load_test_image: changing size to two variables, height and width
